+++ src/sll/api/_generated_raw.c
@@ -1,7 +1,7 @@
 // WARNING: This is an auto-generated file. Any changes made to this file might be lost at any moment. Do Not Edit!
 #include <sll/_sll_internal.h>
 #include <sll/api.h>
-#include <sll/api/_generated.h>
+#include <sll/api/_generated_raw.h>
 #include <sll/array.h>
 #include <sll/common.h>
 #include <sll/ift.h>